Initial commit based on prompt: Help me brainstorm the plot for a game I'm working on. This document is a markdown file which contains a summary of what the game is about, which we'll use later to generate some in-game prompts and dialogues.
The basic plot of the game is that it takes place in the not-too-distant future, once we've got an established colony on mars. We've got tens of thousands of people living on Mars - mostly scientists and engineers as part of the first pass, but now the number of people moving to Mars has stepped up greatly, and there's been a massive influx of people from all walks of life moving there, thanks to the marketing efforts of the world's richest man (who will remain nameless in our game, but is obviously modelled after Elon Musk).

The situation now is that a rebel faction of the colony has splintered off, and a mars civil war is raging. Details about what's happening are hard to come by on Earth, because traditional media has been all but eradicated due to the political situation over the preceding years, and now all news is reported in a highly biased and inaccurate ways by twitter threadbois. All the protagonist of our game knows is that he's been tasked with an urgent mission to deliver highly sensitive medical supplies to the Mars colony, which involves a dangerous run through an asteroid field in the Van Allen belt, which exposes our ship to high levels of radiation. The medical supplies decay rapidly when exposed to this radiation, and our ship comes equipped with a radiation shield generator which protects our cargo during the trip. However, any physical damage to our ship will decrease the effectiveness of our containment field, which leads to the medical supplies expiring more rapidly.

We start with a fixed number of supplies and the goal is to make it to the other side of our space-based obstacle course with the maximum number of viable medical supplies left in our hold. Along the way we face a number of dangers - we must avoid asteroids of course, but it appears that the splinter colonists have made our job even more difficult by adding space mines, turrets, and movable energy barriers that we have to dodge and destroy on our way to make our delivery.

As we make our deliveries, we learn more information from the people we interact with along the way. The story we get back home is that the richest man in the world, who now lives on Mars, is a great benevolent benefactor who's working tirelessly to save humanity. However, the more we learn, the more it becomes clear that he's actually the leader of the splinter colony, after an all-consuming paranoid fear of the "woke mind virus" lead him to believe that the scientists and engineers who worked with him to establish the mars base were actually working against him, and spreading their evil philosophies of acceptance and inclusiveness to undermine his efforts. The scientists and engineers, of course, are mostly concerned with the daily operations of keeping things running and keeping everyone alive and healthy, and they are confused about why their work is being maligned.

In the end, we end up having to decide which side we make our deliveries to, and we'll keep a running leaderboard which tracks which players are contributing the most to each side of the conflict.