Initial commit based on prompt: Let's create a new janus element called audio-factory. On create, we'll dynamically load the ToneJS library from https://unpkg.com/tone and append it to our document's head.
Then once the script is loaded, we'll create an instance of the ToneJS library. We'll have a createSounds() function which will generate a number of different sounds, for now we'll start by creating a simple laser sound named "laserbeam". It'll be a basic "pew" noise that'll play every time the user or a drone fires a laser, probably something simple like a sin wave generator with some filters applied. The sound should be approximately 100ms long.

We'll want to use the existing audiocontext from the JanusXR engine, which we can get by calling

const listener = this.engine.systems.sound.getRealListener()

and looking at listener.context to get the active audio context.