Initial commit based on prompt: Let's create a new element called spacezone-store. This store will load a json file from assets/store-items.json, which contains a number of items, sorted by category, for example:
{
  "Engines": [
    { "name": "Impeller v1", "price": 2500.0, "params": { "thrust": 1200 } },
    { "name": "Impeller v2", "price": 5500.0, "params": { "thrust": 1750 } },
    { "name": "Fusion Drive v1", "price": 15500.0, "params": { "thrust": 3550 } }
  ],
  "Shields": [
    { "name": "Defender v1", "price": 1450.0, "description": "The basic model engine, the Defender v1 provides adequate power to get your ship where it needs to go", "params": { "shieldstrength": 120 } },
    { "name": "Defender v2", "price": 2560.0, "description": "A slight improvement over the v1, the Defender v2 will give you a little extra push to get where you're going a little faster", "params": { "shieldstrength": 160 } }
  ],
  "Cargo": [
    { "name": "Small Cargo Hold", "price": 525, "params": { "cargocapacity": 120 } },
    { "name": "Medium Cargo Hold", "price": 1025, "params": { "cargocapacity": 180 } }
  ]
}

Our element will have a .showItems() function which will display a tabbed view of all of the categories, and a grid of items for sale in each category. Selecting an item from the grid will show the item's full details below, along with a buy button. When the buy button is clicked, we'll emit a "purchased" event, with the data attribute for our event containing the item we just bought.

All CSS needed for this should be placed in a comment at the top of this file.